useComment hook created, useMovie and useUser refactored, components changed to use useComment hook, unit and e2e test fixes
