+++ kinology_frontend/src/components/About/About.test.jsx
@@ -11,8 +11,8 @@ test("renders About component properly", () => {
     </MemoryRouter>
   );
 
-  expect(screen.getByText(/About me/i)).toBeDefined();
-  expect(screen.getByText(/my favorite movies/i)).toBeDefined();
+  expect(screen.getByText(/About author/i)).toBeDefined();
+  expect(screen.getByText(/Author's favorite movies/i)).toBeDefined();
   expect(screen.getByText(/Web app uses TMDB api/i)).toBeDefined();
   expect(screen.getByText(/Fullstack open/i)).toBeDefined();
