+++ kinology_frontend/src/hooks/useUser.js
@@ -2,12 +2,10 @@ import { useEffect, useState, useRef } from "react";
 import { useNavigate } from "react-router-dom";
 import { useNotificationDispatch } from "../contexts/NotificationContext";
 import usersService from "../services/users";
-import commentsService from "../services/comments";
 
 const useUser = (currentUser, removeUser, id) => {
   const [user, setUser] = useState("");
   const [avatar, setAvatar] = useState("");
-  const [comments, setComments] = useState([]);
   const [favoriteMovies, setFavoriteMovies] = useState([]);
   const [watchedMovies, setWatchedMovies] = useState([]);
   const [watchLaterMovies, setWatchLaterMovies] = useState([]);
@@ -26,14 +24,6 @@ const useUser = (currentUser, removeUser, id) => {
       setWatchLaterMovies(fetchedUser.watchLaterMovies);
     };
     fetchUser();
-  }, [id]);
-
-  useEffect(() => {
-    const fetchComments = async () => {
-      const fetchedComments = await commentsService.getComments(id, "profile");
-      setComments(fetchedComments);
-    };
-    fetchComments();
   }, [id]);
 
   const updateUser = async (formData) => {
@@ -60,103 +50,6 @@ const useUser = (currentUser, removeUser, id) => {
         type: "SHOW",
         payload: {
           message: "Something went wrong when updating user",
-          type: "error",
-        },
-      });
-      setTimeout(() => dispatch({ type: "HIDE" }), 5000);
-    }
-  };
-
-  const createComment = async (comment) => {
-    try {
-      commentFormRef.current.toggleVisibility();
-      const createdComment = await commentsService.createComment(
-        id,
-        comment,
-        currentUser,
-        "profile"
-      );
-      setComments(comments.concat(createdComment));
-      dispatch({
-        type: "SHOW",
-        payload: {
-          message: `Comment '${comment}' successfully created`,
-          type: "success",
-        },
-      });
-      setTimeout(() => dispatch({ type: "HIDE" }), 5000);
-    } catch (exception) {
-      dispatch({
-        type: "SHOW",
-        payload: {
-          message: "Something went wrong when creating a comment",
-          type: "error",
-        },
-      });
-      setTimeout(() => dispatch({ type: "HIDE" }), 5000);
-    }
-  };
-
-  // Adding authorId here as well
-  const deleteComment = async (commentId, authorId) => {
-    if (window.confirm("Are you sure you want to delete the comment?")) {
-      try {
-        const filteredComments = comments.filter((c) => c.id !== commentId);
-        setComments(filteredComments);
-        await commentsService.deleteComment(
-          user.id,
-          commentId,
-          currentUser,
-          authorId,
-          "profile"
-        );
-        dispatch({
-          type: "SHOW",
-          payload: {
-            message: `Comment successfully deleted`,
-            type: "success",
-          },
-        });
-        setTimeout(() => dispatch({ type: "HIDE" }), 5000);
-      } catch (exception) {
-        dispatch({
-          type: "SHOW",
-          payload: {
-            message: "Something went wrong when deleting a comment",
-            type: "error",
-          },
-        });
-        setTimeout(() => dispatch({ type: "HIDE" }), 5000);
-      }
-    }
-  };
-
-  const updateComment = async (commentId, comment, authorId) => {
-    try {
-      const updatedComment = await commentsService.updateComment(
-        id,
-        commentId,
-        currentUser,
-        comment,
-        authorId,
-        "profile"
-      );
-      setComments(
-        comments.map((c) => (c.id === updatedComment.id ? updatedComment : c))
-      );
-      dispatch({
-        type: "SHOW",
-        payload: {
-          message: `Comment successfully updated with '${comment}'`,
-          type: "success",
-        },
-      });
-      setTimeout(() => dispatch({ type: "HIDE" }), 5000);
-    } catch (exception) {
-      dispatch({
-        type: "SHOW",
-        payload: {
-          message: `Something went wrong when updating a comment`,
           type: "error",
         },
       });
@@ -201,16 +94,12 @@ const useUser = (currentUser, removeUser, id) => {
     user,
     avatar,
     setAvatar,
-    comments,
     favoriteMovies,
     watchedMovies,
     watchLaterMovies,
     updateFormRef,
     commentFormRef,
     updateUser,
-    createComment,
-    deleteComment,
-    updateComment,
     deleteUser,
   };
 };
